+++ app/build.gradle
@@ -24,6 +24,9 @@ android {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
+        debug {
+            debuggable true
+        }
     }
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_19
@@ -33,7 +36,8 @@ android {
         jvmTarget = '19'
     }
     buildFeatures {
-        viewBinding=true
+        viewBinding = true
+        buildConfig = true
     }
 }
