+++ app/src/main/AndroidManifest.xml
@@ -26,8 +26,15 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
-        <!--        <activity android:name="site.pnpl.mira.ui.MainActivity"/>-->
+        <provider
+            android:name=".utils.MiraFileProvider"
+            android:authorities="site.pnpl.mira.provider"
+            android:exported="false"
+            android:grantUriPermissions="true">
+            <meta-data
+                android:name="android.support.FILE_PROVIDER_PATHS"
+                android:resource="@xml/provider_paths"/>
+        </provider>
     </application>
 
 </manifest>
