+++ app/src/main/java/site/pnpl/mira/di/AppComponent.kt
@@ -9,8 +9,9 @@ import site.pnpl.mira.ui.greeting.fragments.AcquaintanceFragment
 import site.pnpl.mira.ui.greeting.fragments.GreetingFragment
 import site.pnpl.mira.ui.greeting.fragments.InterNameFragment
 import site.pnpl.mira.ui.greeting.fragments.SplashFragment
-import site.pnpl.mira.ui.home.HomeFragment
+import site.pnpl.mira.ui.home.fragments.HomeFragment
 import site.pnpl.mira.ui.home.HomeViewModel
+import site.pnpl.mira.ui.home.fragments.SettingsFragment
 import javax.inject.Singleton
 
 @Singleton
@@ -29,5 +30,6 @@ interface AppComponent {
     fun inject(exercisesListViewModel: ExercisesListViewModel)
     fun inject(checkInViewModel: CheckInViewModel)
     fun inject(homeFragment: HomeFragment)
+    fun inject(settingsFragment: SettingsFragment)
 
 }
