+++ app/src/main/java/site/pnpl/mira/ui/home/recycler_view/ItemTouchHelperCallback.kt
@@ -3,7 +3,7 @@ package site.pnpl.mira.ui.home.recycler_view
 import android.graphics.Canvas
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.RecyclerView
-import site.pnpl.mira.ui.home.HomeFragment
+import site.pnpl.mira.ui.home.fragments.HomeFragment
 
 class ItemTouchHelperCallback(
     private val onChangeExpandedListener: HomeFragment.ChangeExpandedListener
