+++ app/src/main/java/site/pnpl/mira/utils/MiraFileProvider.kt
@@ -1,0 +1,5 @@
+package site.pnpl.mira.utils
+
+import androidx.core.content.FileProvider
+
+class MiraFileProvider : FileProvider()
