+++ app/src/main/res/anim/check_in_up.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <set xmlns:android="http://schemas.android.com/apk/res/android"
-    android:duration="500">
+    android:duration="300">
     <translate
         android:fromYDelta="100%"
         android:toYDelta="0"/>
