+++ app/src/main/res/layout/fragment_element.xml
@@ -13,6 +13,7 @@
         android:layout_height="0dp"
         app:lottie_rawRes="@raw/cro_gif"
         app:lottie_autoPlay="true"
+        app:lottie_loop="true"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
