+++ app/src/main/res/layout/fragment_greeting.xml
@@ -25,10 +25,11 @@
 
     <com.airbnb.lottie.LottieAnimationView
         android:id="@+id/animation"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
+        android:layout_width="300dp"
+        android:layout_height="300dp"
         app:lottie_autoPlay="true"
-        app:lottie_rawRes="@raw/mira"
+        app:lottie_loop="true"
+        app:lottie_rawRes="@raw/with_glass"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
