+++ app/src/main/res/values/strings.xml
@@ -96,5 +96,13 @@
     <string name="label_hi">Привет</string>
     <string name="label_info_1">"Пока что здесь пусто\nНачни записывать свои ощущения ежедневно, это поможет разобраться в себе"</string>
     <string name="label_info_2">"За выбранный период нет записей\nВыбери другие даты, используя календарь"</string>
+    <string name="tittle_settings">Настройки</string>
+    <string name="button_save_change">Сохранить изменения</string>
+    <string name="about_project"><u>О проекте</u></string>
+    <string name="share"><u>Поделиться</u></string>
+    <string name="button_change_saved">Изменения сохранены</string>
+    <string name="calendar_tittle">Начало - Конец</string>
+    <string name="calendar_positive_button">Показать</string>
+    <string name="calendar_negative_button">Отмена</string>
 
 </resources>
