+++ app/src/main/res/xml/provider_paths.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<paths>
+    <files-path name="assets" path="."/>
+    <cache-path name="images" path="/" />
+</paths>
